// src/models/Product.ts
interface Product {
    alcoholLevel: number;
    aroma: any;
    associatedProductList: any;
    availableForThePartner: boolean;
    availableInTheMarket: boolean;
    bcCode: any;
    brand: any;
    brandQuality: any;
    category: any;
    color: any;
    consumptionTips: string;
    countryOfOrigin: string;
    description: string;
    ebxLastUpdateDate: string;
    foodPairing: any;
    gs1: any;
    gtinCodeList: any[];
    gtrExclusive: boolean;
    id: string;
    ingredientList: string;
    language: any;
    localFields: any;
    manufacturerDetails: string;
    medalsAndAwards: any;
    medalsAndAwardsList: any[];
    mediaList: any[];
    nutritional: any;
    objectID: string;
    onlineExclusive: boolean;
    packagingDetails: string;
    packshot: any;
    partnerActivationDate: any;
    path: string;
    pimUpdateDate: number;
    productHistory: string;
    productName: string;
    productType: string;
    retailer: any;
    size: any;
    skuStatus: string;
    taste: any;
    tastingNotes: any;
    translatedInTheLocale: any;
    translationActivatedOn: any;
    vap: any;
    variety: any;
    vineyardHistory: any;
    vintage: any;
}

export default Product;